[nrf noup] riscv: Fix test_k_sleep
On haltium FPGA RISCV the overhead introduced by the k_sem_take()
operation is around 50ms, given that in the test the thread_sleep is
already taking ~50ms a timeout of 100ms total on the semaphore is
definitely too tight.

Increase the timeout to 100ms to have a bigger margin when waiting
on the semaphore (a timeout of 200ms should be enough when
considering 100ms spent in the thread and 50ms overhead).

To be investigated why the k_sem_take() operation is so slow but my
assumption is that going through ecall / mret for each context switch is
having a bigger impact than anticipated.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>
(cherry picked from commit f14be28d30572430462815612940c7131d5e32d9)
(cherry picked from commit c30f17427e90a21fe0bb906bb50c05df09a78020)