[nrf noup] drivers: ieee802154_nrf5: EUI64 from BLE.ADDR for nRF54H
In order to get a to some extent unique EUI64 for nRF54H BLE.ADDR
was used in the IEEE 802.15.4 driver to create such an identifier.

This is possibly a temporary solution, as the amount of EUI64-s
available in such a way is very limited (~16 million). However
currently there does not seem to be another feasible way to
get a device identifiers (such are kept in SICR, to which the
radio core has no access.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>
(cherry picked from commit 337ec838788c5c28282b9ce6a52f370b72d633a6)