[nrf noup] tests: timer_behavior: Use TCM RAM on APP core for nRF54H20 soc1
The one_tick_timer_train test case is sensible to execution speed and
it turns out that the access time to global RAM used without data cache
(which cannot be enabled in the soc1 revision due to a hardware issue)
is too high for this test case to pass.
Use the TCM RAM on the APP core in the timer_behavior test then.

Ref. NRFX-3455

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>